Refactor code for consistency and readability
- Adjusted spacing and formatting in various classes including DIYRecipeButton, ScrollButton, JeiDIYRecipeCategory, and others for improved readability.
- Updated comments for clarity in ScrollButton and Gradient classes.
- Ensured consistent use of parentheses in conditional statements across multiple files.
- Removed unnecessary imports and streamlined code in MultiblockHandler and ShapeUtils.
- Enhanced error handling in MultiblockItem and WallMultiblockItem constructors.
- Improved method documentation for better understanding of parameters and functionality.

+++ src/main/java/tanukidecor/block/entity/PhonographBlockEntity.java
@@ -10,10 +10,7 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
-import net.minecraft.sounds.SoundEvent;
-import net.minecraft.sounds.SoundEvents;
 import net.minecraft.stats.Stats;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.world.ContainerHelper;
@@ -26,21 +23,16 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.RecordItem;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.JukeboxBlock;
 import net.minecraft.world.level.block.LevelEvent;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.gameevent.GameEvent;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.Vec3;
-import tanukidecor.TDRegistry;
 import tanukidecor.block.RotatingTallBlock;
 import tanukidecor.block.misc.PhonographBlock;
 
 import javax.annotation.Nullable;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
 
 public class PhonographBlockEntity extends SingleSlotBlockEntity {
 
@@ -53,7 +45,7 @@ public class PhonographBlockEntity extends SingleSlotBlockEntity {
         super(pType, pPos, pBlockState);
     }
 
-    //// METHODS ////
+    /// / METHODS ////
 
     public static void tick(Level level, BlockPos blockPos, BlockState blockState, PhonographBlockEntity blockEntity) {
         blockEntity.phonographTick(level, blockPos, blockState);
@@ -61,7 +53,7 @@ public class PhonographBlockEntity extends SingleSlotBlockEntity {
 
     public static InteractionResult use(BlockState blockState, Level level, BlockPos pos, Player player, InteractionHand hand, BlockHitResult hitResult) {
         // validate block entity
-        if(!(level.getBlockEntity(pos) instanceof PhonographBlockEntity blockEntity)) {
+        if (!(level.getBlockEntity(pos) instanceof PhonographBlockEntity blockEntity)) {
             return InteractionResult.FAIL;
         }
         // validate side
@@ -69,14 +61,14 @@ public class PhonographBlockEntity extends SingleSlotBlockEntity {
             return InteractionResult.SUCCESS;
         }
         // drop recording, if any
-        if(!blockEntity.isEmpty()) {
+        if (!blockEntity.isEmpty()) {
             ItemStack record = blockEntity.getFirstItem();
             // remove item
             blockEntity.clearContent();
             // give to player
-            if(player.getItemInHand(hand).isEmpty()) {
+            if (player.getItemInHand(hand).isEmpty()) {
                 player.setItemInHand(hand, record);
-            } else if(!player.getInventory().add(record)) {
+            } else if (!player.getInventory().add(record)) {
                 Block.popResource(level, pos.above(), record);
             }
             // stop playing sound
@@ -86,7 +78,7 @@ public class PhonographBlockEntity extends SingleSlotBlockEntity {
         }
         // insert recording, if any
         final ItemStack itemStack = player.getItemInHand(hand);
-        if(blockEntity.isEmpty() && !itemStack.isEmpty() && blockEntity.canPlaceItem(0, itemStack)) {
+        if (blockEntity.isEmpty() && !itemStack.isEmpty() && blockEntity.canPlaceItem(0, itemStack)) {
             // update item stack (sound is handled in setChanged)
             ItemStack record = itemStack.split(blockEntity.getMaxStackSize());
             blockEntity.setFirstItem(record);
@@ -101,7 +93,7 @@ public class PhonographBlockEntity extends SingleSlotBlockEntity {
     public static void onRemove(BlockState blockState, Level level, BlockPos pos, BlockState newState, boolean isMoving) {
         if (level.getBlockEntity(pos) instanceof PhonographBlockEntity blockEntity) {
             // drop contents and stop music
-            if(!blockEntity.isEmpty()) {
+            if (!blockEntity.isEmpty()) {
                 level.levelEvent(LevelEvent.SOUND_STOP_JUKEBOX_SONG, pos, 0);
                 blockEntity.dropContents();
             }
@@ -143,13 +135,13 @@ public class PhonographBlockEntity extends SingleSlotBlockEntity {
         if (pLevel instanceof ServerLevel serverlevel) {
             Direction direction = getBlockState().getValue(RotatingTallBlock.FACING);
             Vec3 vec3 = Vec3.atCenterOf(pPos).add(0.4D * direction.getStepX(), -0.125D, 0.4D * direction.getStepZ());
-            float f = (float)pLevel.getRandom().nextInt(4) / 24.0F;
-            serverlevel.sendParticles(ParticleTypes.NOTE, vec3.x(), vec3.y(), vec3.z(), 0, (double)f, 0.0D, 0.0D, 1.0D);
+            float f = (float) pLevel.getRandom().nextInt(4) / 24.0F;
+            serverlevel.sendParticles(ParticleTypes.NOTE, vec3.x(), vec3.y(), vec3.z(), 0, f, 0.0D, 0.0D, 1.0D);
         }
 
     }
 
-    //// CONTAINER ////
+    /// / CONTAINER ////
 
     @Override
     public void setChanged() {
@@ -194,7 +186,7 @@ public class PhonographBlockEntity extends SingleSlotBlockEntity {
         this.setItem(0, pRecord);
     }
 
-    //// RECORD ////
+    /// / RECORD ////
 
     protected void startPlaying() {
         this.recordStartedTick = this.tickCount;
@@ -223,14 +215,14 @@ public class PhonographBlockEntity extends SingleSlotBlockEntity {
         return !this.getFirstItem().isEmpty() && this.isPlaying;
     }
 
-    //// CLEARABLE ////
+    /// / CLEARABLE ////
 
     @Override
     public void clearContent() {
         this.inventory.clear();
     }
 
-    //// NBT ////
+    /// / NBT ////
 
     private static final String KEY_IS_PLAYING = "IsPlaying";
     private static final String KEY_RECORD_STARTED_TICK = "RecordStartTick";
